pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "userbank"
version = "1.0.0"
authors = [
  { name="Jordi Carrera Ventura", email="jordi.carrera.ventura@gmail.com" },
]
description = "A Pythonic wrapper for Google Sheets based user management."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "PyYAML",
    "python-dotenv",
]

[project.scripts]
userbank = "userbank.cli:main"

[project.urls]
"Google page" = "https://github.com/JordiCarreraVentura/userbank"
